main: split text protocol into proto_text.c
export a lot of the connection handling code from memcached.c
